architectural rework - moved models into dedicated files
