+++ lettings/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
-from .models import Letting
-from .models import Address
+from .models.letting import Letting
+from .models.address import Address
 
 
 admin.site.register(Letting)
