+++ lettings/models/address.py
@@ -34,18 +34,3 @@ class Address(models.Model):
 
     def __str__(self):
         return f'{self.number} {self.street}'
-
-
-class Letting(models.Model):
-    """
-    Represents a rental listing.
-
-    Attributes:
-        title (str): Title of the rental listing.
-        address (Address): Associated address of the rental.
-    """
-    title = models.CharField(max_length=256)
-    address = models.OneToOneField(Address, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.title
