+++ profiles/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Profile
+from .models.profile import Profile
 
 
 admin.site.register(Profile)
