+++ profiles/tests/test_profiles_views.py
@@ -5,7 +5,7 @@ Integration tests for the 'profiles' views.
 import pytest
 from django.urls import reverse
 from django.contrib.auth.models import User
-from profiles.models import Profile
+from profiles.models.profile import Profile
 
 
 @pytest.mark.django_db
